requirements.txt
google-genai>=1.38.0
streamlit>=1.49.1
mailerlite-python